Move OpenRouter routing values out of legacy settings
