+++ src/puripuly_heart/config/settings.py
@@ -48,6 +48,10 @@ from puripuly_heart.config.settings_vnext.schema import (  # noqa: F401
     AppSettingsVNext,
     new_anonymous_telemetry_identifier,
 )
+from puripuly_heart.core.openrouter_routing import (
+    OpenRouterProviderRouting,
+    OpenRouterRoutingMode,
+)
 
 SETTINGS_SCHEMA_VERSION = 25
 MANAGED_AUTH_CLAIM_SOURCE_DISCORD = "discord"
@@ -256,21 +260,6 @@ class OpenRouterLLMModel(str, Enum):
     DEEPSEEK_V4_FLASH = OPENROUTER_MODEL_DEEPSEEK_V4_FLASH
     GEMINI_37_FLASH = OPENROUTER_MODEL_GEMINI_37_FLASH
     GEMINI_31_FLASH_LITE = OPENROUTER_MODEL_GEMINI_31_FLASH_LITE
-
-
-class OpenRouterRoutingMode(str, Enum):
-    LATENCY = "latency"
-
-
-class OpenRouterProviderRouting(str, Enum):
-    DEFAULT = "default"
-    DEEPSEEK_ONLY = "deepseek_only"
-    GOOGLE_GEMINI_LATENCY = "google_gemini_latency"
-    GEMMA4_26B_31B_LATENCY = "gemma4_26b_31b_latency"
-    GEMMA4_31B_LATENCY = "gemma4_31b_latency"
-    GEMMA4_26B_LATENCY = "gemma4_26b_latency"
-    DEEPSEEK_V4_FLASH_LATENCY = "deepseek_v4_flash_latency"
-    GEMMA4_31B_CEREBRAS_ONLY = "gemma4_31b_cerebras_only"
 
 
 class OpenRouterCredentialSource(str, Enum):
