+++ tests/architecture/test_dependency_boundaries.py
@@ -375,14 +375,6 @@ KNOWN_ALLOWED_VIOLATIONS: frozenset[ImportViolation] = frozenset(
             reason="runtime owners must coordinate through domain events, resolved DTOs, lifecycle/message/observability protocols, not app wiring, Flet UI, provider config parsing, or concrete adapters",
         ),
         ImportViolation(
-            rule_id="providers-avoid-ui-settings-and-runtime-log-concretes",
-            importer="src/puripuly_heart/providers/llm/openrouter.py",
-            imported="puripuly_heart.config.settings",
-            importer_layer="providers",
-            imported_layer="migration/serialization",
-            reason="providers may use provider ports, SDKs, and message/observability protocols, but not Flet UI, settings migration internals, app services, or concrete SessionRuntimeLoggingService-style adapters",
-        ),
-        ImportViolation(
             rule_id="ui-adapters-avoid-provider-construction",
             importer="src/puripuly_heart/ui/desktop_overlay.py",
             imported="puripuly_heart.config.settings",
@@ -1731,6 +1723,11 @@ def test_r00_legacy_settings_reachability_census_matches_the_pinned_baseline() -
 
     expected_current = {entry["path"]: set(entry["symbols"]) for entry in entries}
     expected_current.pop("src/puripuly_heart/main.py")
+    expected_current.pop("src/puripuly_heart/providers/llm/openrouter.py")
+    expected_current["src/puripuly_heart/app/wiring/wiring_llm_factory.py"] -= {
+        "OpenRouterProviderRouting",
+        "OpenRouterRoutingMode",
+    }
     expected_current["src/puripuly_heart/app/services/canonical_settings_persistence.py"] = {
         "AppSettings"
     }
@@ -1772,9 +1769,9 @@ def test_r00_retires_stable_profile_import_and_secret_copy_surfaces() -> None:
     assert occurrences == set()
 
 
-def test_a01_provider_observation_boundary_reduces_dependency_debt_to_25() -> None:
-    assert len(KNOWN_ALLOWED_VIOLATIONS) == 25
-    assert len(_dependency_violations()) == 25
+def test_a02_openrouter_settings_boundary_reduces_dependency_debt_to_24() -> None:
+    assert len(KNOWN_ALLOWED_VIOLATIONS) == 24
+    assert len(_dependency_violations()) == 24
 
 
 def test_r00_canonical_migration_loader_has_no_flat_ingress() -> None:
